+++ src/icemac/addressbook/browser/person/tests/test_address.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+import pytest
+
+
 def test_address__menu__1(person_data, browser):
     """The entries in the add menu have a specific sort order."""
     browser.login('editor')
@@ -146,9 +149,10 @@ def test_address__DeletePostalAddressForm__3(person_data, browser):
         browser.getControl('main postal address').displayOptions)
 
 
-def test_address__DeletePostalAddressForm__4(person_data, browser):
-    """`DeletePostalAddressForm` cannot be accessed by a visitor."""
-    browser.login('visitor')
+@pytest.mark.parametrize('loginname', ['visitor', 'archivist'])
+def test_address__DeletePostalAddressForm__4(person_data, browser, loginname):
+    """It cannot be accessed by some roles."""
+    browser.login(loginname)
     browser.assert_forbidden(browser.POSTAL_ADDRESS_DELETE_URL)
 
 
@@ -166,9 +170,10 @@ def test_address__DeletePhoneNumberForm__1(person_data, browser):
             browser.getControl('main phone number').displayOptions)
 
 
-def test_address__DeletePhoneNumberForm__2(person_data, browser):
-    """`DeletePhoneNumberForm` cannot be accessed by a visitor."""
-    browser.login('visitor')
+@pytest.mark.parametrize('loginname', ['visitor', 'archivist'])
+def test_address__DeletePhoneNumberForm__2(person_data, browser, loginname):
+    """It cannot be accessed by some roles."""
+    browser.login(loginname)
     browser.assert_forbidden(browser.PHONE_NUMBER_DELETE_URL)
 
 
@@ -186,9 +191,10 @@ def test_address__DeleteEMailAddressForm__1(person_data, browser):
             browser.getControl('main e-mail address').displayOptions)
 
 
-def test_address__DeleteEMailAddressForm__2(person_data, browser):
-    """`DeleteEMailAddressForm` cannot be accessed by a visitor."""
-    browser.login('visitor')
+@pytest.mark.parametrize('loginname', ['visitor', 'archivist'])
+def test_address__DeleteEMailAddressForm__2(person_data, browser, loginname):
+    """It cannot be accessed by some roles."""
+    browser.login(loginname)
     browser.assert_forbidden(browser.EMAIL_ADDRESS_DELETE_URL)
 
 
@@ -206,7 +212,9 @@ def test_address__DeleteHomePageAddressForm__1(person_data, browser):
             browser.getControl('main home page address').displayOptions)
 
 
-def test_address__DeleteHomePageAddressForm__2(person_data, browser):
-    """`DeleteHomePageAddressForm` cannot be accessed by a visitor."""
-    browser.login('visitor')
+@pytest.mark.parametrize('loginname', ['visitor', 'archivist'])
+def test_address__DeleteHomePageAddressForm__2(
+        person_data, browser, loginname):
+    """It cannot be accessed by some roles."""
+    browser.login(loginname)
     browser.assert_forbidden(browser.HOMEPAGE_ADDRESS_DELETE_URL)