+++ src/icemac/addressbook/browser/search/tests/test_multi_keyword.py
@@ -9,7 +9,8 @@ def test_multi_keyword__Search__1(search_data, browser):
         ['Select requested keywords from the list popping up when selecting'
          ' the keywords control.'] ==
         browser.etree.xpath(
-            '//div[@id="content"]/div[@class="row no-print"]/text()'))
+            '//div[@id="content"]//div[@class="row no-print explanation"]'
+            '/text()'))
 
 
 def test_multi_keyword__Search__2(address_book, browser):
@@ -18,7 +19,7 @@ def test_multi_keyword__Search__2(address_book, browser):
     browser.open(browser.SEARCH_BY_KEYWORD_URL)
     browser.getControl('Search').click()
     assert ('No person found.' == browser.etree.xpath(
-        '//div[@id="content"]/div[@class="row"]/text()')[-1].strip())
+        '//div[@id="content"]//div[@class="row result"]/text()')[-1].strip())
 
 
 def test_multi_keyword__Search__3(search_data, browser):
@@ -26,7 +27,7 @@ def test_multi_keyword__Search__3(search_data, browser):
     browser.login('visitor')
     browser.keyword_search('work')
     assert ('No person found.' == browser.etree.xpath(
-        '//div[@id="content"]/div[@class="row"]/text()')[-1].strip())
+        '//div[@id="content"]//div[@class="row result"]/text()')[-1].strip())
 
 
 def test_multi_keyword__Search__4(search_data, browser):